Split refreshrs into separate component
